Add an undocumented DS Lite firmware version, comment out the switch case
That's right, I own one of these undocumented ones.

Also, this switch case is unreliable, firmwares seem to have data
outside the range of flash chips available. So to err on the side of
caution, let's just save the whole 512KiB, regardless of whether the
chip actually has that capacity.